+++ imports/ui/blast/BlastResultPlot.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ContainerDimensions from 'react-container-dimensions'
 //import { Popover, OverlayTrigger } from 'react-bootstrap';
-import { scaleLinear } from 'd3-scale';
-import { interpolateGreys } from 'd3-scale-chromatic';
+import { scaleLinear, interpolateGreys } from 'd3';//-scale';
+//import { interpolateGreys } from 'd3-scale-chromatic';
 
 const PopoverHover = (props) => {
   return (