+++ imports/startup/client/index.js
@@ -1,4 +1,6 @@
-import './client-fixtures.js';
-import './routes.jsx';
 import '/imports/ui/global_stylesheets/global.scss';
 import '/imports/ui/global_stylesheets/fontello/css/fontello.css';
+
+import './client-fixtures.js';
+import './load-fonts.js';
+import './routes.jsx';