+++ imports/startup/client/load-fonts.js
@@ -1,0 +1,7 @@
+import WebFont from 'webfontloader';
+
+WebFont.load({
+  google: {
+    families: ['Roboto:300']
+  }
+})